Tidied up old files
