--- OutsideESP32/I2S.cpp
@@ -1,89 +1,0 @@
-// Modified from https://github.com/MhageGH/esp32_SoundRecorder/tree/master
-// by MhageGH
-
-#include "I2S.h"
-
-void i2s_init() {
-  i2s_config_t i2s_config = {
-    .mode = (i2s_mode_t)(I2S_MODE_MASTER | I2S_MODE_RX),
-    .sample_rate = I2S_SAMPLE_RATE,
-    .bits_per_sample = i2s_bits_per_sample_t(I2S_SAMPLE_BITS),
-    .channel_format = I2S_CHANNEL_FMT_ONLY_LEFT,
-    .communication_format = (i2s_comm_format_t)(I2S_COMM_FORMAT_I2S | I2S_COMM_FORMAT_I2S_MSB),
-    .intr_alloc_flags = 0,
-    .dma_buf_count = 4,
-    .dma_buf_len = 1024,
-    .use_apll = false
-  };
-
-  i2s_driver_install(I2S_NUM_0, &i2s_config, 0, NULL);
-
-  const i2s_pin_config_t pin_config = {
-    .bck_io_num = I2S_SCK,
-    .ws_io_num = I2S_WS,
-    .data_out_num = -1,
-    .data_in_num = I2S_SD
-  };
-
-  i2s_set_pin(I2S_PORT, &pin_config);
-
-}
-
-void i2s_adc_data_scale(uint8_t * d_buff, uint8_t* s_buff, uint32_t len) 
-{
-    uint32_t j = 0;
-    uint32_t dac_value = 0;
-    for (int i = 0; i < len; i += 2) {
-        dac_value = ((((uint16_t) (s_buff[i + 1] & 0xf) << 8) | ((s_buff[i + 0]))));
-        d_buff[j++] = 0;
-        d_buff[j++] = dac_value * 256 / 2048;
-    }
-}
-
-void create_wav_header(byte* header, int wav_file_size) {
-  header[0] = 'R';
-  header[1] = 'I';
-  header[2] = 'F';
-  header[3] = 'F';
-  unsigned int file_size_minus8 = wav_file_size + 44 - 8;
-  header[4] = (byte)(file_size_minus8 & 0xFF);
-  header[5] = (byte)((file_size_minus8 >> 8) & 0xFF);
-  header[6] = (byte)((file_size_minus8 >> 16) & 0xFF);
-  header[7] = (byte)((file_size_minus8 >> 24) & 0xFF);
-  header[8] = 'W';
-  header[9] = 'A';
-  header[10] = 'V';
-  header[11] = 'E';
-  header[12] = 'f';
-  header[13] = 'm';
-  header[14] = 't';
-  header[15] = ' ';
-  header[16] = 0x10;
-  header[17] = 0x00;
-  header[18] = 0x00;
-  header[19] = 0x00;
-  header[20] = 0x01;
-  header[21] = 0x00;
-  header[22] = 0x01;
-  header[23] = 0x00;
-  header[24] = 0x80;
-  header[25] = 0x3E;
-  header[26] = 0x00;
-  header[27] = 0x00;
-  header[28] = 0x00;
-  header[29] = 0x7D;
-  header[30] = 0x00;
-  header[31] = 0x00;
-  header[32] = 0x02;
-  header[33] = 0x00;
-  header[34] = 0x10;
-  header[35] = 0x00;
-  header[36] = 'd';
-  header[37] = 'a';
-  header[38] = 't';
-  header[39] = 'a';
-  header[40] = (byte)(file_size_minus8 & 0xFF);
-  header[41] = (byte)((file_size_minus8 >> 8) & 0xFF);
-  header[42] = (byte)((file_size_minus8 >> 16) & 0xFF);
-  header[43] = (byte)((file_size_minus8 >> 24) & 0xFF);
-}
