+++ old_esp32files/esp_32cam_send_photo/esp_32cam_send_photo.ino
