+++ old_pyscripts/mqtt_camera.py
